Add different executor modes including virtualenv #10
This means the Poe the Poet now also works well without poetry 🎉 I'll need
to think of another tagline 🤔.

Four executors are available: poetry, virtualenv, simple (inherits env from
shell), and auto which attempts to choose sensibly between the other three in
the stated order of preference based on the contents of pyproject.toml and the
parent directory.

PoeExecutor implementations register themselves much like task classes so it
should be pretty easy to add more if needed.

There's also a new global config option at `tool.poe.executor.type` for
specifying which executor to use, which default to auto. If this is set to
"virtualenv" then `tool.poe.executor.location` may also be provided to specify
a relative path to the virtualenv to use.

Also refactored various internals to better accomodate the new executor API.