Always explicitly use the poetry virtualenv if possible in the poetry executor
This removes an optimization that probably does add much value anymore, but
does seem to cause tasks to be run without the correct virtual_env in some
scenarios on windows.

Also make process of updaating the path env var to enable a venv try to avoid
redundant updates in case the virtualenv is already active.