Allow use of the original working directory for tasks
For some tasks, it may be useful to use the working directory from which
the task has been executed, since they may be meant to operate on some
files that the user points them to. This commit introduces a special
'$exec_cwd' value for the 'cwd' task option, which does exactly that.
Existing use-cases are unaffected by this addition and work as usual.