Preserve '--' pseudo argument when passing to tasks
This commit makes it so that when a user passes the `--` pseudo-argument
to a task, it is preserved for further processing by the task itself.
Various commands may use `--` to e.g. pass arguments to another
subprogram, or (e.g. in the case of `argparse`) to differentiate
positional arguments from options.

Note that this change potentially breaks compatibility, as some already
implemented tasks may rely on the `--` pseudo-argument being dropped, in
fact, one of the tests was written that way. However, I strongly believe
this is actually a bug and I decided to update the test as well.