Work with _OLD_VIRTUAL_PATH env var like a venv activate script
Also set _OLD_VIRTUAL_PYTHONHOME like in an activate script.

This makes the virtualenv activation behaviour of the poetry and virtualenv
executors behave in a way closer to the standard.

Fix a test that previously relied on accidental PATH inheritance from the
root project venv, which is no longer possible.