fix: improve detection of poetry 2 projects (#274)
Projects created with poetry >= 2.0 don't necessarily include a
tool.poetry table in the pyproject.toml. Thus poethepoet cannot
rely on the presence or absence of this table to tell whether
it is appropriate to default to using the PoetryExecutor to run
tasks with poetry's venv.

It seems the best remaining indicator is the content of the
build-system which poetry still configures consistently, although
this is less ideal.