fix: Don't resolve symlinks to uv in UvExecutor (#283)
`uv` is installed via Snap on Ubuntu. Snap is a wrapper providing a safe environment for execution.

Snap symlinks all executable commands to it's binary.

* Called as `snap`: Snap CLI
* Called as a symlink from any other file: Execute the file in it's environment

The `UvExecutor` class resolved the path itself including any symlinks and eventually calls `/usr/bin/snap` instead of `/snap/bin/uv`. Snaps assumes a call to it's own CLI and fails.

With this patch, the shell will resolve the path and provide the correct "executed file" information (`$0` in a Bash script) to Snap.

fixes #282