feat!: Improve handling of unmatched glob patterns in cmd tasks (#315)
Unmatched glob patterns will now be passed through to the task untouched,
instead of being replaced with nothing.

The old behavior can be restored for a task be setting the new `empty_glob` 
option to "null" (like the nullglob option in bash). This option can also 
be set to "fail" to make the task execution fail with an error when an 
unmatched glob pattern is encountered.

In case a task fails in a way that might have been due to this breaking
change then we write a warning message with a link to explain the breaking 
change.

Also:
- resolve relative paths for task level cwd early in the process.